requirements.txt
django
djangorestframework
pymongo
django-cors-headers
python-dotenv
nltk
PyJWT
requests
beautifulsoup4